Buat admin export pdf, delete account, final touch

+++ delete.php
@@ -8,6 +8,8 @@ if (isset($_GET["id"])) {
     $table = $_GET["table"];
     $category = $_GET["data"];
 
+    $user_details = query("SELECT * FROM users WHERE id_user = $id_post");
+
     // var_dump($id_post);
     // var_dump($header);
     // var_dump($table);
@@ -17,8 +19,6 @@ if (isset($_GET["id"])) {
     if ($header == "index" or $header == "category") {
         $query = "DELETE FROM $table WHERE id_post = $id_post";
         mysqli_query($connect, $query);
-        // header("Location:" . $header . ".php?category=" . $category);
-
         if ($category == "null") {
             echo "<script>
             alert('Post deleted successfully');
@@ -33,18 +33,23 @@ if (isset($_GET["id"])) {
     } else if ($header == "comment") {
         $query = "DELETE FROM $table WHERE id_comment = $id_post";
         mysqli_query($connect, $query);
-        // header("Location:" . $header . ".php?id=" . $category);
         echo "<script>
             alert('Comment deleted successfully');
             document.location.href = 'comment.php?id=$category';
         </script>";
     } else if ($header == "profile") {
-        $query = "DELETE FROM $table WHERE id_user = $id_post";
-        mysqli_query($connect, $query);
-        // header("Location:" . $header . ".php?id=" . $category);
-        echo "<script>
+        if ($user_details[0]["status"] == "admin") {
+            echo "<script>
+            alert('Admin cannot be deleted');
+            document.location.href = 'profile.php?id=$id_post';
+        </script>";
+        } else {
+            $query = "DELETE FROM $table WHERE id_user = $id_post";
+            mysqli_query($connect, $query);
+            echo "<script>
             alert('User deleted successfully');
             document.location.href = 'index.php';
         </script>";
+        }
     }
 }
